Move services to util
